+++ packages/address/src/index.ts
@@ -1,4 +1,5 @@
 export * from '@shopify/address-consts';
 
 export * from './loader';
+export * from './format';
 export {default} from './AddressFormatter';